# apps/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Applad API
  version: 1.0.0
  description: Self-hosted Backend-as-a-Service API
  license:
    name: BSD-3-Clause
servers:
  - url: /v1
    description: Default (relative)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Applad-Key
  parameters:
    projectHeader:
      name: X-Applad-Project
      in: header
      required: true
      schema:
        type: string
...
paths:
...
  /flags/{flagId}:
    get:
      summary: Get a flag
      tags: [Flags]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
        - name: flagId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Flag details
    put:
      summary: Update a flag
      tags: [Flags]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
        - name: flagId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Flag updated
    delete:
      summary: Delete a flag
      tags: [Flags]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
        - name: flagId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...